+++ packages/analytics/dashboard-renderer/sandbox/pages/RendererDemo.vue
@@ -43,13 +43,13 @@ const context: DashboardRendererContext = {
     type: 'relative',
     time_range: '24h',
   },
-  refreshInterval: 5000,
+  refreshInterval: 0,
 }
 
 const dashboardConfig: DashboardConfig = {
   gridSize: {
     cols: 6,
-    rows: 9,
+    rows: 5,
   },
   tileHeight: 167,
   tiles: [
@@ -74,11 +74,11 @@ const dashboardConfig: DashboardConfig = {
     {
       definition: {
         chart: {
-          type: ChartTypes.HorizontalBar,
-          chartTitle: 'Horizontal bar chart of mock data',
-          allowCsvExport: true,
+          type: ChartTypes.TopN,
+          chartTitle: 'Top N chart of mock data',
+          description: 'Description',
         },
-        query: {},
+        query: { limit: 1 },
       },
       layout: {
         position: {
@@ -87,7 +87,8 @@ const dashboardConfig: DashboardConfig = {
         },
         size: {
           cols: 3,
-          rows: 2,
+          rows: 1,
+          fitToContent: true,
         },
       },
     } as TileConfig,
@@ -98,7 +99,7 @@ const dashboardConfig: DashboardConfig = {
           chartTitle: 'Top N chart of mock data',
           description: 'Description',
         },
-        query: {},
+        query: { limit: 3 },
       },
       layout: {
         position: {
@@ -107,24 +108,24 @@ const dashboardConfig: DashboardConfig = {
         },
         size: {
           cols: 3,
-          rows: 2,
+          rows: 1,
+          fitToContent: true,
         },
       },
     } as TileConfig,
     {
       definition: {
         chart: {
-          type: ChartTypes.TopN,
-          chartTitle: 'Top N chart of mock data',
-          description: 'Description',
-          entityLink: `https://test.com/cp/${CP_ID_TOKEN}/entity/${ENTITY_ID_TOKEN}`,
+          type: ChartTypes.HorizontalBar,
+          chartTitle: 'Horizontal bar chart of mock data',
+          allowCsvExport: true,
         },
         query: { dimensions: ['route'] },
       },
       layout: {
         position: {
           col: 0,
-          row: 3,
+          row: 2,
         },
         size: {
           cols: 3,
@@ -145,7 +146,7 @@ const dashboardConfig: DashboardConfig = {
       layout: {
         position: {
           col: 3,
-          row: 3,
+          row: 2,
         },
         size: {
           cols: 3,
@@ -166,7 +167,7 @@ const dashboardConfig: DashboardConfig = {
       layout: {
         position: {
           col: 0,
-          row: 5,
+          row: 4,
         },
         size: {
           cols: 1,
@@ -185,7 +186,7 @@ const dashboardConfig: DashboardConfig = {
       layout: {
         position: {
           col: 1,
-          row: 5,
+          row: 4,
         },
         size: {
           cols: 1,
@@ -204,7 +205,7 @@ const dashboardConfig: DashboardConfig = {
       layout: {
         position: {
           col: 2,
-          row: 5,
+          row: 4,
         },
         size: {
           cols: 3,